newcon/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
python-docx==1.1.2
docx2txt==0.8
PyMuPDF==1.21.1
python-pptx==0.6.21
fpdf==1.7.2
Pillow==9.2.0
pdfkit==1.0.0
reportlab==3.6.12
flask-cors==5.0.1